Made naming more compliant with PEP, moved main to a new file
